Introduce storage model for reservation instead of migration.
