+++ src/main/scala/mesosphere/marathon/core/instance/Instance.scala
@@ -407,7 +407,7 @@ object Instance {
 
   implicit val instanceStateFormat: Format[InstanceState] = Json.format[InstanceState]
 
-  implicit val reservationFormat: Format[Reservation] = Reservation.reservationFormat
+  //  implicit val reservationFormat: Format[Reservation] = Reservation.reservationFormat
 
   implicit lazy val tasksMapFormat: Format[Map[Task.Id, Task]] = Format(
     Reads.of[Map[String, Task]].map {
