+++ src/main/scala/mesosphere/marathon/storage/migration/MigrationTo18100.scala
@@ -7,12 +7,12 @@ import akka.stream.scaladsl.Flow
 import com.typesafe.scalalogging.StrictLogging
 import mesosphere.marathon.api.v2.json.Formats
 import mesosphere.marathon.core.condition.Condition
-import mesosphere.marathon.core.instance.{Goal, Reservation}
+import mesosphere.marathon.core.instance.Goal
 import mesosphere.marathon.core.instance.Instance.{AgentInfo, Id, InstanceState, agentFormat}
 import mesosphere.marathon.core.storage.store.PersistenceStore
 import mesosphere.marathon.core.task.{Task, TaskCondition}
 import mesosphere.marathon.core.task.state.NetworkInfo
-import mesosphere.marathon.state.{Instance, Timestamp}
+import mesosphere.marathon.state.{Instance, Reservation, Timestamp}
 import mesosphere.marathon.storage.repository.InstanceRepository
 import play.api.libs.json._
 import play.api.libs.json.Reads._
